cloud/transcription validation/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-iot
pycrypto